brain/requirements.txt
fastapi
uvicorn
supabase
google-genai
python-dotenv
pydantic
python-multipart
pdf2image
pillow